removes switchable_nr_plugin
