+++ src/game/mod.rs
@@ -3,4 +3,3 @@ pub mod one_holding_switchable_nr_couple_plugin;
 pub mod path_plugin;
 pub mod pause_plugin;
 pub mod status_bar_plugin;
-pub mod switchable_nr_plugin;
